\name{wang}
\docType{data}
\alias{wang}
\title{Two-person DNA mixture published by Wang et al. (2006)}
\description{
  Two-person DNA mixture typed on 13 STR loci. 
}
\usage{
## The data is stored in a comma separated spread sheet file (.csv).
## The location of the file is returned by running the code below:
## cat(paste(.find.package("mixsep"),"data",sep=.Platform$file.sep))
}
\format{A spread sheet with three columns: Locus, Allele and Area.}
\references{
 T Wang, N Xue and JD Birdwell (2006). 'Least-Square Deconvolution: A
 Framework for Interpreting Short Tandem Repeat Mixtures'. Journal of
 Forensic Science 51 (6): 1284-1297.
}
