\name{makeShortcut}
\alias{makeShortcut}
\concept{Genetics}
\concept{Deconvolution}
\concept{Forensic Genetics}
\title{Make Desktop shortcut on Windows operating system}
\description{
  Creates a shortcut on the desktop - Windows only.
}
\usage{makeShortcut()}
\value{
  Creates a shortcut on the desktop. The function identifies the
  location of the mixsep-package in the file system and adds a
  .bat-file, which launches R in the terminal. This is the best and
  easiet way of using the mixture separator.
}
\examples{
## run this in Windows to create shortcut on the desktop. 
\dontrun{makeShortcut()}
}

